BE (tests) - abandon csv file on edition creation
